requirements.txt
Pillow==10.2.0
streamlit==1.32.0 